Move new TDirectory code to a shared place.
